+++ R/calc_light_merged.R
@@ -23,21 +23,19 @@
 #' @export
 #' @examples 
 #' \dontrun{
-#' library(mda.streams)
 #' library(dplyr)
-#' library(lubridate)
 #' library(ggplot2)
 #' library(unitted)
-#' coords <- get_site_coords('nwis_08062500')
-#' PAR.obs <- get_ts('par_calcSw', 'nwis_08062500') %>%
-#'   mutate(solar.time = convert_UTC_to_solartime(DateTime, coords$lon)) %>%
-#'     select(solar.time, light=par) %>%
-#'     subset(solar.time %within% interval(ymd("2008-03-10"), ymd("2008-03-14"), tz=tz(solar.time)) &
-#'           !(solar.time %within% interval(as.POSIXct("2008-03-12 10:00", tz=tz(solar.time)), 
-#'                                          as.POSIXct("2008-03-12 14:00", tz=tz(solar.time)))))
-#' PAR.mod <- u(data.frame(solar.time=seq(as.POSIXct('2008-03-11', tz=tz(PAR.obs$solar.time)),
-#'     as.POSIXct('2008-03-15', tz=tz(PAR.obs$solar.time)), by=as.difftime(10, units='mins'))) %>%
-#'   mutate(light=calc_light(u(solar.time), latitude=coords$lat, longitude=coords$lon)))
+#' timebounds <- as.POSIXct(c('2008-03-12 00:00', '2008-03-12 23:59'), tz='UTC')
+#' coords <- list(lat=32.4, lon=-96.5)
+#' PAR.obs <- data_frame(
+#'   solar.time=seq(timebounds[1], timebounds[2], by=as.difftime(3, units='hours')),
+#'   light=c(0, 0, 85.9, 1160.5, 1539.0, 933.9, 0, 0)
+#' ) %>% as.data.frame()
+#' PAR.mod <- data_frame(
+#'   solar.time=seq(timebounds[1], timebounds[2], by=as.difftime(0.25, units='hours')),
+#'   light=calc_light(solar.time, latitude=coords$lat, longitude=coords$lon)
+#' ) %>% as.data.frame()
 #' PAR.merged <- calc_light_merged(PAR.obs, PAR.mod$solar.time, 
 #'   latitude=coords$lat, longitude=coords$lon, max.gap=as.difftime(20, units='hours'))
 #' ggplot(bind_rows(mutate(v(PAR.obs), type='obs'), mutate(v(PAR.mod), type='mod'), 
@@ -50,6 +48,22 @@ calc_light_merged <- function(
   PAR.obs=mm_data(solar.time, light), 
   solar.time, latitude, longitude, max.PAR=NA,
   max.gap=as.difftime(3, units="hours"), attach.units=is.unitted(PAR.obs)) {
+  
+  # ensure units are correct and present within this function
+  arg_units <- list(
+    PAR.obs=get_units(mm_data(solar.time, light)),
+    solar.time='',
+    latitude='degN',
+    longitude='degE')
+  for(argname in names(arg_units)) {
+    arg <- get(argname)
+    unit <- arg_units[[argname]]
+    if(is.unitted(arg)) {
+      stopifnot(all(get_units(arg) == unit))
+    } else {
+      assign(argname, u(arg, unit))
+    }
+  }
   
   . <- is.mod <- obs <- mod <- resid.abs.int <- resid.prop.int <- merged <- '.dplyr.var'
   
@@ -94,26 +108,33 @@ calc_light_merged <- function(
   }
   
   # create a 'merged' time series where modeled values are adjusted to flow
-  # through observed values
+  # through observed values.
+  
+  # compute the modeled values and then the residuals as both differences and 
+  # proportions, dealing with 0 specially
   PAR.merged <- PAR.merged %>%
     mutate(
-      # model light
       mod = calc_light(solar.time, latitude, longitude, max.PAR),
-      # compute the residuals as both differences and proportions, dealing with NA and 0 specially
-      resid.abs = ifelse(is.na(obs), 0, obs-mod),
-      resid.prop = ifelse(mod==u(0, 'umol m^-2 s^-1') | (is.na(obs) & mod==u(0, 'umol m^-2 s^-1')), 1, obs/mod))
-  # interpolate the residuals to match up with every modeled light value. pipes fail with this approx call, so use boring notation
-  PAR.merged$resid.abs.int <- approx(x=PAR.merged$solar.time, y=PAR.merged$resid.abs, xout=PAR.merged$solar.time, rule=2)$y
-  if(is.unitted(PAR.merged)) PAR.merged$resid.abs.int <- u(PAR.merged$resid.abs.int, get_units(PAR.merged$obs))
-  PAR.merged$resid.prop.int <- approx(x=PAR.merged$solar.time, y=PAR.merged$resid.prop, xout=PAR.merged$solar.time, rule=2)$y
-  PAR.merged <- PAR.merged %>%
-    # do the correction from mod scale to obs scale. purely absolute or purely proportional can give some funky values, so use the 
-    mutate(merged = u(ifelse(resid.prop.int <= 1, mod * resid.prop.int, mod + resid.abs.int), get_units(mod)))
+      resid.abs = obs - mod,
+      resid.prop = ifelse(mod==u(0, 'umol m^-2 s^-1'), NA, obs / mod))
   
+  # interpolate the residuals to match up with every modeled light value. pipes 
+  # fail with this approx call, so use boring notation
+  PAR.obsonly <- PAR.merged[!is.na(PAR.merged$obs), ]
+  PAR.merged$resid.abs.int <- approx(x=PAR.obsonly$solar.time, y=v(PAR.obsonly$resid.abs), xout=v(PAR.merged$solar.time), rule=2)$y
+  PAR.merged$resid.prop.int <- approx(x=PAR.obsonly$solar.time, y=PAR.obsonly$resid.prop, xout=PAR.merged$solar.time, rule=2)$y
+  
+  # do the correction from mod scale to obs scale
   PAR.merged <- PAR.merged %>%
-    # collect just the rows and cols we want
+    mutate(
+      merged = u(ifelse(resid.prop.int <= 1, mod * resid.prop.int, pmax(0, v(mod) + resid.abs.int)), get_units(mod)))
+  
+  # collect just the rows and cols we want
+  PAR.merged <- PAR.merged %>%
     subset(is.mod) %>%
     select(solar.time, light=merged)
+  
+  if(!attach.units) PAR.merged <- v(PAR.merged)
   
   # return
   PAR.merged